101-natural.c answer

+++ 0x02-functions_nested_loops/101-natural.c
@@ -10,11 +10,11 @@
 int main(void)
 {
 	int i;
-	int result = 0;
+	long int result = 0;
 
 	for (i = 1; i < 1024; i++)
 		if ((i % 3 == 0) || (i % 5 == 0))
 			result = result + i;
-	printf("%d \n", result);
+	printf("%ld \n", result);
 	return (0);
 }
